requirements.txt
pygpgme==0.3
python-gnupg==0.3.9
coloredlogs==5.0
plant==0.1.2
PTable==0.9.2
cryptography==1.6
pycrypto==2.6.1
envelop==0.2.0
pyqrcode==1.2.1
